+++ packages/plugins/src/plugins/baskt/types.ts
@@ -1,6 +1,0 @@
-import { ID } from '../../utils/sui/types/id';
-
-export type Receipt = {
-  id: ID;
-  amountDeposited: string;
-};