+++ packages/tx-parser/src/utils/parseTransaction/getRelevantInstructions.ts
@@ -5,18 +5,31 @@ import {
 } from '@solana/web3.js';
 import { isParsedInstruction } from './isParsedInstruction';
 
+const memoProgramIds = [
+  'MemoSq4gqABAXKb96qnH8TysNcWxMyWCqXgDLGmfcHr',
+  'Memo1UhkJRfHyvLMcVucJwxXeuD728EqVDDwQDxFMNo',
+];
+
 export const getRelevantInstructions = (txn: ParsedTransactionWithMeta) =>
   txn.transaction.message.instructions
     .map((i) => {
+      if (
+        i.programId.toString() === 'ComputeBudget111111111111111111111111111111'
+      ) {
+        return null;
+      }
+
+      if (memoProgramIds.includes(i.programId.toString())) {
+        return null;
+      }
+
       if (isParsedInstruction(i)) {
-        return i.programId.toString() !==
-          'ComputeBudget111111111111111111111111111111' &&
-          !(
-            i.programId.toString() === '11111111111111111111111111111111' &&
-            i.parsed.type === 'advanceNonce'
-          )
-          ? i
-          : null;
+        if (
+          i.programId.toString() === '11111111111111111111111111111111' &&
+          i.parsed.type === 'advanceNonce'
+        ) {
+          return null;
+        }
       }
       return i;
     })