+++ packages/core/CHANGELOG.md
@@ -2,6 +2,42 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.15.7](https://github.com/sonarwatch/portfolio/compare/core-0.15.6...core-0.15.7) (2025-05-19)
+
+
+
+## [0.15.6](https://github.com/sonarwatch/portfolio/compare/core-0.15.5...core-0.15.6) (2025-05-18)
+
+
+
+## [0.15.5](https://github.com/sonarwatch/portfolio/compare/core-0.15.4...core-0.15.5) (2025-05-16)
+
+
+
+## [0.15.4](https://github.com/sonarwatch/portfolio/compare/core-0.15.3...core-0.15.4) (2025-05-14)
+
+
+
+## [0.15.3](https://github.com/sonarwatch/portfolio/compare/core-0.15.2...core-0.15.3) (2025-05-14)
+
+
+
+## [0.15.2](https://github.com/sonarwatch/portfolio/compare/core-0.15.1...core-0.15.2) (2025-05-14)
+
+
+
+## [0.15.1](https://github.com/sonarwatch/portfolio/compare/core-0.15.0...core-0.15.1) (2025-05-14)
+
+
+
+# [0.15.0](https://github.com/sonarwatch/portfolio/compare/core-0.14.205...core-0.15.0) (2025-05-14)
+
+
+
+## [0.14.205](https://github.com/sonarwatch/portfolio/compare/core-0.14.204...core-0.14.205) (2025-05-14)
+
+
+
 ## [0.14.204](https://github.com/sonarwatch/portfolio/compare/core-0.14.203...core-0.14.204) (2025-05-13)